requirements.txt
attrs==22.1.0
black==22.10.0
browserOpen==1.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dropbox==11.36.0
exceptiongroup==1.0.4
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
ply==3.11
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
stone==3.3.1
tomli==2.0.1
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.12
uWSGI==2.0.21
vcrpy==4.2.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
